Pod_retriever/requirements.txt
tensorflow==2.17.0
numpy
Pillow